% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.r
\name{dgp_scale}
\alias{dgp_scale}
\title{Generate scaled bivariate normal data compared to a bivariate normal distribution
with unit marginal variances.}
\usage{
dgp_scale(n, mu, Sigma, v = c(1, 1))
}
\arguments{
\item{n}{sample size.}

\item{mu}{expectation, a vector of size two.}

\item{Sigma}{covariance matrix.}

\item{v}{scaling vector.}
}
\value{
a \eqn{n\times 2} matrix of the data.
}
\description{
Generate scaled bivariate normal data compared to a bivariate normal distribution
with unit marginal variances.
}
\author{
Nadja Klein.
}
