% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.r
\name{dgp_trans}
\alias{dgp_trans}
\title{Generate translated bivariate normal data (with a certain mean and covariance matrix) compared to a zero mean distribution.}
\usage{
dgp_trans(n, mu, Sigma, v = c(0, 0))
}
\arguments{
\item{n}{sample size.}

\item{mu}{expectation, a vector of size two.}

\item{Sigma}{covariance matrix.}

\item{v}{translation vector.}
}
\value{
a \eqn{n\times 2} matrix of the data.
}
\description{
Generate translated bivariate normal data (with a certain mean and covariance matrix) compared to a zero mean distribution.
}
\author{
Nadja Klein.
}
