% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomqu.r
\name{find_good_alpha}
\alias{find_good_alpha}
\title{On Affine Equivariant Multivariate Quantiles}
\usage{
find_good_alpha(X, eps = 0.025)
}
\arguments{
\item{X}{must be a (n,2) matrix.}

\item{eps}{stopping criterion.}
}
\value{
a list with a the selected rows of the data, X is the X(a) matrix, X_reduced is the data matrix X without the selected columns.
}
\description{
This function returns the transformation matrix X(a), see Chakraborty, B. (2001).
}
\references{
Chakraborty, B. (2001). On affine equivariant multivariate quantiles. 
\emph{Annals of the Institute of Statistical Mathematics}, \bold{53}, 380--403. <https://doi.org/10.1023/A:1012478908041>.
}
\seealso{
\code{\link{geomqu}}.
}
\author{
Nadja Klein, Paul Wiemann.
}
