% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.bivquant}
\alias{plot.bivquant}
\title{Plot the estimated bivariate quantiles (for the  CDF (cumulative distrubtion function)-based quantiles) and the data for a fitted bivquant object. See ?bivquant for more details.}
\usage{
\method{plot}{bivquant}(x,
  mains = c("CDF-based quantiles (original scale)",
  "CDF-based quantile (unit square)"), ...)
}
\arguments{
\item{x}{a fitted \code{bivquant} model.}

\item{mains}{title of figures.}

\item{...}{further arguments for function \code{\link{plot}}.}
}
\description{
Plot the estimated bivariate quantiles (for the  CDF (cumulative distrubtion function)-based quantiles) and the data for a fitted bivquant object. See ?bivquant for more details.
}
\seealso{
\code{\link{plot}} for details.
}
\author{
Nadja Klein.
}
