% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.geomqu}
\alias{plot.geomqu}
\title{Plot the estimated bivariate quantiles (for the geometric quantiles of Chakraborty, B. (2001).) and the data. See ?geomqu for more details.}
\usage{
\method{plot}{geomqu}(x, ...)
}
\arguments{
\item{x}{a fitted \code{geomqu} model.}

\item{...}{further arguments for function \code{\link{plot}}.}
}
\description{
Plot the estimated bivariate quantiles (for the geometric quantiles of Chakraborty, B. (2001).) and the data. See ?geomqu for more details.
}
\seealso{
\code{\link{plot}} for details.
}
\author{
Nadja Klein, Paul Wiemann.
}
